Add manifest & build scripts (#1)
This PR adds all the setup necessary to run MapLibre build jobs in Jenkins. Since this is the first PR to this repo, the `build-android.sh` script, which builds the library, won't be run. The jobs in this PR just get the Android environment set up in a docker container. I'll create a follow-up PR to get the actual build started.

- Add [Dockerfile.cirunner](https://github.com/lyft/maplibre-gl-native-private/pull/1/files#diff-31a1b93408c82dacaedc04b4a50d181510b43d8796cc86f56ed838bd49d20118) to set up android environment (mostly taken from [AV Ride Guide](https://github.com/lyft/avrideguideandroidci/blob/main/docker/Dockerfile.cisetup) project)
- Add [manifest.yaml](https://github.com/lyft/maplibre-gl-native-private/pull/1/files#diff-d865b06b497efc8c56c63e5bedfc86f6da6a005cdb7d30e3702286f3d918aaa6) to configure jenkins job
- Add [build-android.sh](https://github.com/lyft/maplibre-gl-native-private/pull/1/files#diff-d8d60744938bb045a45b3ce53f370ff1ed2a776b131ceba691da81a03ca4469d) which will run the maplibre android build
- Delete a bunch of MapLibre CI scripts that were causing Lyft CI tasks to fail